feat(form-field): add state-layer to input-wrapper
This commit adds a new div with class "state-layer" to the input-wrapper in form-field.component.html. It also adds styles for this new element in form-field.component.scss, including setting its position to absolute and visibility to hidden by default. When the user hovers over an enabled input field, the background color of the state-layer changes to match that of the input-wrapper.